changed entity structure by splitting it at files
